+++ .idea/androidTestResultsUserPreferences.xml
@@ -75,6 +75,7 @@
                 <map>
                   <entry key="Duration" value="90" />
                   <entry key="Google Pixel 5" value="120" />
+                  <entry key="Pixel_3a_API_33_x86_64" value="120" />
                   <entry key="Tests" value="360" />
                 </map>
               </option>
