+++ core/androidTest/src/main/kotlin/me/jerryokafor/ihenkiri/android/test/app/screens/LandingScreenTest.kt
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package me.jerryokafor.ihenkiri.android.test.app
+package me.jerryokafor.ihenkiri.android.test.app.screens
 
 import androidx.compose.ui.test.assertHasClickAction
 import androidx.compose.ui.test.assertIsDisplayed
